#165 add separate expandable menu for protocols
